ueba-fraud/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
pydantic==2.8.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
Faker==25.8.0
requests==2.32.3
